moved things to gamestage component
